parallel-orchestration/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
agent-framework --pre
agent-framework-devui --pre
azure-identity